Debug prismatic joint + default offset
Prismatic joint is now working with a PID controller (command emitted
by main controller which is a bit weird but works for now). Also, the
sim now starts with an initial rotation using the command-line option
in the gazebo.launch file. Both are in service of simulating the foot
fall of the robot.