Make a pendulum that swings on impact
Strangely, the choice of joints in the robot seems to impact what
the landing behavior is very strongly. In a separate stash (to be
committed), the robot bounces on impact, whereas here we get the
expected behavior of the arm swinging due to the momentum of the
motor and wheel.

This situation is much more challenging for the controller, which
performed confusingly well in the other scenario; now we know it's
because the momentum somehow disappears.