Make custom node for prismatic joint controller
Implement a custom controller for the prismatic joint, implementing the spring
within as well. This controller for now applies a constant force whenever the
joint is expanding to increase jump height over time.

Note the joint controller rate needed to be updated because the existing 50Hz
was too slow for the spring physics to converge. I don't know if we really
_need_ 1kHz but it does (seem to) work at that rate at least.

Also refactor the contact detection to its own separate node now that two
controllers are using it.