Implement hopping with a pull solenoid
Pull on the solenoid as the robot lifts off the ground to compress a
spring, let go when the robot is at the bottom of its bounce. Surprisingly
the energy transfer is not very efficient and the bounce seems to get lost
in the default frictions of the system. To my understanding this is already
among the best timings for a pull solenoid.

A push solenoid transfers more energy by foregoing the energy storage spring
altogether, and of course you can make one by just flipping one of the links
around. Let's try that next.