Use solenoid on push stroke, move to top of pendulum
Using a separate spring to store the solenoid energy is inefficient,
because the solenoid force is non-linear with length, so any spring
would either be too stiff near equilibrium, or too soft at max
compression. This is detailed in the previous commit too. Now, the
solenoid uses a push stroke (to be achieved physically by flipping
one of the interfaces to the links upside down and using a pull stroke)
and is able to deliver more energy.

Also, @lkunelk 's suggestion fixed the energy loss issue from the
inelastic collisions at the solenoid and spring end stops by putting
all the mass on one end, so the systems before and after the collision
don't differ much in mass and energy is mostly preserved.