Make all principal inertias non-zero
Turns out Gazebo just really doesn't like it when the inertias are
zero. I wanted them that way so the dummy links I'm putting in to,
say, join a revolute and prismatic joint together don't affect the
simulation results. However, turns out setting them to zero does
the exact opposite and ruins the entire sim. In the first case a
few commits ago, the zero inertia on the base link was preventing
the whole robot from rotating when it hit the ground, which
happened to be fixed by adding a fixed joint to a dummy link
before the prismatic spring joint. I had thought it was a problem
with the revolute+prismatic combo somehow, but it recurred here
when adding the solenoid's prismatic joint. Now I can get rid of
all the random links that exist to put the fixed joints that
happened to stabilize the sim before. Gazebo/ODE should really be
the one here to tell you that zero inertias will pulverize your
simulation. I could understand if the result was as described in
https://wiki.ros.org/urdf/Tutorials/Adding%20Physical%20and%20Collision%20Properties%20to%20a%20URDF%20Model:

> When using realtime controllers, inertia elements of zero (or
almost zero) can cause the robot model to collapse without warning,
and all links will appear with their origins coinciding with the
world origin.

But instead we saw a robot that acted like it had infinite inertia
which is the absolute opposite of what you would expect. I had to
reconstruct the entire robot link-by-link in a separate simulation
to find this out by the way. Don't be deceived by the miniscule
changes. This was not fun.