Add global constraints to robot
Force the robot into the YZ plane to reduce the instabilities of landing
on a corner of the base link. Does this defeat the purpose of using
a 3D simulator? Temporarily, yes. These are training wheels, since the
energy of the solenoid is just enough to overcome the losses baked into
the system. Note that these joints also seem to take energy out of the
system, so the simulation as a whole needs some evaluation from an
energetic standpoint.