Better estimate landing angle, re-enable path tracking
Previously, we were using the position of the base link to estimate the
trajectory of the robot, but as the robot repositioned to land, the
link would swivel and threw off the estimate. The ideal solution is to
track the center of mass, and now that almost all the mass is at the
top of the robot, we can just track the flywheel. However, the flywheel
isn't the part that's hitting the ground, so we have to compensate.

The equations coming from conservation of energy originally made the
landing velocity very easy to calculate, but having to compensate for
the angle of the arm, which is itself dependent on the landing velocity,
makes the equation not fun analytically. Instead, it turns out that we
can just run it as a recurrence relation, because the way it is done in
`main.cpp` always overestimates the landing height and is hence stable.
In practice, it converges very quickly, to within 3 significant digits
in a couple of iterations.

Now the robot is bouncing everywhere, it's also important to refactor
the base link. To help things out a bit, this commit makes the base
link into a proper foot with a wide and short box, and gives it a very
high inertia in all axes so it stays locked to its initial world
orientation. This removes the need for the world constraint joints
which caused problems keeping the energy inside the robot before.

Housekeeping note: joint and link state message accesses are being
modified to index by searching the name in the `names` fields
instead of using fixed numeric indices. These were coming in
unreliable order, seemingly because of the addition of the tracking
dot that made a race condition for which robot was loaded first. This
commit begins util and const files to house these housekeeping items.