Begin custom worldfile and physics
Set CFM and update rate at launch time for Gazebo. Previously, the
robot was sliding (and violating momentum conservation) because the
contact was unstable due to the many prismatic things happening in the
arm. With softer constraints, the friction is more accurate at the
default physics update step. Hopefully this doesn't take too much
energy out of the system.

Other changes:

- Default physics rate set to 50 (prev. 1000)
- Camera changed to look at the robot head-on from the front (+Y)