Remove name from base link collision
Friction (and kp) was being stripped from base_link during parsing because
of a bug in gazebo's URDF->SDF conversion. See osrf/gazebo#2497 . Fixing the
base link friction should eliminate the need for CFM/KP which was initially
put in to improve the collision detection with the world and stop the
robot from slipping during the pushoff phase.