Analyze and tune solenoid
These changes are part of experiments to understand the sensitivities
of the solenoid. They revealed for the DSTL-0829 solenoid we're using
that the initial force plateau means that the energy transfer is
entirely through the speed of the body's mass, because the spring
won't compress under this force profile. This is alarming, as the
only useful part of the solenoid force is whatever it manages to
generate at maximum extension. Unfortunately, precompressing the
solenoid actually doesn't lead to significant gains even with the
optimal spring (e.g. one whose K follows the slope of the solenoid
force curve).

It's worth now exploring other options for the hopping joint.