Model 2-link piston vs. linearized prismatic joint
Previously we used a prismatic joint linearized around the piston angle=0
(maximum torque->force transfer). However, the results were suggesting
choices with huge amounts of expansion (e.g. >10cm) which is much larger
than the servo-driven piston ranges we're considering. This improves the
simulation accuracy by modelling the two links of the piston directly.
Verified with the current model in the `dev-servo-piston` branch.