Use XOR on hall sensors instead of timer loop
The config for the timer interaction between TIM1 and TIM4 needed a bit
of futzing beyong what CUBE provided to get both the hall edge-to-edge delay
and the COM preloading/interrupt behavior, which can be seen in the user code
section in main.

- TIM4 capture 1 and output compare 2 need to be enabled (fair enough)
- TIM4 commutation delay (i.e. CCR2) needs to be >0 to trigger COM
  (unexpected: the default CUBE delay is 0): this is because otherwise
  the reset trigger won't generate a posedge for TIM1 to catch
- TIM1 CR2 needs to enable both CCUS and CCPC to trigger COM on the
  tim_trgi input (coming from TIM4's CCR2)
- TIM1 DIER needs to explicitly enable COM interrupt (for some reason
  I thought the relevant interrupts were enabled via HAL_TIM_Base_Start
  and were somehow stored in the htim instance).

Nothing too surprising in retrospect but fair to say a least a bit confusing.