First attempt at control
Rudimentary PID and filtering in the motor loop. Currently, control updates
are in the hall-sensor callback, but this doesn't make sense, given the control
loop speed is then tied to motor speed. The open-loop timer callback on TIM6 has
been revived to deal with standstill - this is where the control loop will
probably end up living.